PublicApiGenerator made source only package
Bumped version to 5.0
CecilEx.cs moved into ApiGenerator
Renamed PublicApiGenerator to ApiGenerator to not clash with namespace
