+++ src/PublicApiGenerator/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Mono.Cecil" version="0.9.6.1" targetFramework="net40" />
+  <package id="Mono.Cecil" version="0.9.6.4" targetFramework="net40" />
 </packages>
